packages/processors/pyproject.toml
[tool.poetry]
name = "processors"
version = "0.1.0"
description = ""
authors = ["Mark Shteyn <markshteyn1@gmail.com>"]
readme = "README.md"
packages = [{ include = "processing_utils" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
requests = "^2.32.3"
pypdf2 = "^3.0.1"
pymupdf = "^1.25.1"
python-pptx = "^0.6.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
